Note View Status not being included in solutions
When inactivate views are added to a solution the exported the solution does not contain the statecode for the view. 
statecode	Edm.Int32	Shows the status of the view.
Display Name: Status
Default Options 
Value  Label
0         Active
1         Inactive

Therefore when an Inactive views are imported into the target they are imported as Active by default.