Update software-requirements-for-microsoft-dynamics-365-server.md
Support regularly fields questions and requests for documentation on how to host D365 CE on a Azure Managed SQL instance. Our stance is that this is not currently supported for D365 CE on premise deployments. The purpose of this document change is to (hopefully) deflect the question before it comes to support and offer an easy reference for support to quickly deflect the ticket.